Connect Posts and Users

+++ REST-Service-Blog/REST-API-Server/controllers/feed.js
@@ -2,6 +2,7 @@ const fs = require("fs");
 const path = require("path");
 const { validationResult } = require("express-validator");
 const Post = require("../models/post");
+const User = require("../models/user");
 
 //-----------------
 // Controller: getPosts
@@ -67,22 +68,32 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
 
-  console.log("CREATE POST");
+  let creator;
 
   // Create post in db
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Jedo" },
+    creator: req.userId,
   });
 
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post Created Successfully!",
         post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -159,6 +170,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized.");
+        statusCode = 403;
+        throw error;
+      }
+
       // Delete the old image if there's a new image uploaded
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
@@ -200,10 +217,23 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized.");
+        statusCode = 403;
+        throw error;
+      }
+
       // Delete the image
       clearImage(post.imageUrl);
 
       return Post.findByIdAndDelete(postId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then((result) => {
       res.status(200).json({ message: "Post Deleted Successfully." });
